requirements.txt
# Core framework
fastapi==0.115.5
uvicorn[standard]==0.32.1

# Data validation & settings
pydantic==2.9.2
pydantic-settings==2.6.1

# HTTP client (async)
httpx==0.27.2

# OpenAI SDK (async-compatible)
openai==1.54.4

# Pinecone vector database
pinecone-client==5.0.1

# Environment variable loading
python-dotenv==1.0.1
